Rename blocks file.
