+++ python/spinn/util/blocks.py
